music-theory-assistant/requirements.txt
# Core
pandas
scikit-learn
tqdm
python-dotenv

# Qdrant
qdrant-client[fastembed]>=1.14.2

# UI / API
streamlit
fastapi
uvicorn[standard]

# LLMs (OpenAI client)
openai

# (Optional) evaluation
sentence-transformers

# Monitoring
psycopg2-binary>=2.9
prometheus-client